% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cducsuTweets.R
\docType{data}
\name{cducsuTweets}
\alias{cducsuTweets}
\title{cducsuTweets: Example resource for twitter_fas()-output}
\format{
resource (list of list of tibbles)
}
\usage{
data(cducsuTweets)
}
\description{
This dataset contains tweets from the Fraction "CDU/CSU" in the
German Bundestag from Thursday, 11th March 2021. The tweets only
contain public tweets and mentions from persons of public interest.
}
\keyword{datasets}
