% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getTweets}
\alias{getTweets}
\title{Get Dataframe from resource}
\usage{
getTweets(
  .data,
  withUsers = TRUE,
  withRetweeted = TRUE,
  withMedia = TRUE,
  tz = "UTC"
)
}
\arguments{
\item{.data}{Resource, retrieved by \code{twitter_fas()}.}

\item{withUsers}{(Logical) if true, add information about author
and the authors of referred tweets.}

\item{withRetweeted}{(Logical) if true, add information about the retweeted
or cited tweet.}

\item{withMedia}{(Logical) if true, add information about media included
within the tweet.}

\item{tz}{Format dates with time zone}
}
\value{
A Data Frame with Tweet data from Resource
}
\description{
Get Dataframe from resource
}
\examples{
data(cducsuTweets)
tweets <- getTweets(cducsuTweets)
head(tweets)


}
