% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitteRacademic.R
\docType{package}
\name{twitteRacademic}
\alias{twitteRacademic}
\title{twitteRacademic: provide access to twitter academic product track (API V2)}
\description{
The package twitteRacademic is intended for academics who have an academic
account on Twitter-Developer.
}
\details{
This package enables the integration of the Academic Research product track
of Twitter, which provides free access to the full-archive search and other
V2 endpoints. It provides a feature for circumstantial use of the
 full-archive search call, which is used to access the historical API.

The twitteRacademic package provides a function to retrieve information from
twitter full-academic-search endpoint as well as some functions to convert
the resulting resource into a dataframe.
}
\section{twitteRacademic functions}{

The \code{twitter_fas()}-function retrieves Information. The
\code{getTweet()}-function derives a dataset of tweets from the resource,
the \code{get*()}-functions add information about the authors, referred
referred tweets and media in these tweets.
}

