% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{withMedia}
\alias{withMedia}
\title{Add enclosed Media-Information to Tweet-Dataframe}
\usage{
withMedia(res, tweets)
}
\arguments{
\item{res}{A resource, retrieved by \code{twitter_fas()}.}

\item{tweets}{The tweets-dataframe, as returned by \code{getTweets()}}
}
\value{
A data frame, actually with the length of the tweets-data, with media-information added.
}
\description{
Add enclosed Media-Information to Tweet-Dataframe
}
\examples{
tweets <- withMedia(cducsuTweets, getTweets(cducsuTweets))
head(tweets)
}
