% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{withRetweeted}
\alias{withRetweeted}
\title{Add Information on retweeted tweets to Tweet-Dataframe}
\usage{
withRetweeted(res, tweets)
}
\arguments{
\item{res}{A resource, retrieved by \code{twitter_fas()}.}

\item{tweets}{The tweets-dataframe, as returned by \code{getTweets()}}
}
\value{
A data frame, actually with the length of the tweets-data, with information on retweeted data added.
}
\description{
Add Information on retweeted tweets to Tweet-Dataframe
}
\examples{
tweets <- withRetweeted(cducsuTweets, getTweets(cducsuTweets))
head(tweets)


}
