% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter_fas.r
\name{twitter_fas}
\alias{twitter_fas}
\title{twitter_fas - Full Academic Search in Twitter Academic Research Track (API 2.0)}
\usage{
twitter_fas(query, start, end, bearer_token, max_results = 500)
}
\arguments{
\item{query}{String to search for. Check https://developer.twitter.com/en/docs/twitter-api/tweets/search/api-reference/get-tweets-search-all for further details.}

\item{start}{Starting time for full archive research. Formattad as string of form: "2021-01-01T00:00:01Z"}

\item{end}{End time for full archive research. Formattad as string of form: "2021-03-01T00:00:01Z"}

\item{bearer_token}{Developer Token for authorization to Academic Project. An application at developer.twitter.com is required.}

\item{max_results}{The Twitter api retrieves a maximum of 500 tweets per request.
This function  defaults to the maximum of 500, but can be lowered by setting this argument.}
}
\value{
List of Data Frames
}
\description{
twitter_fas - Full Academic Search in Twitter Academic Research Track (API 2.0)
}
\examples{
source("~/.Rprofile")
bearer_token <- twitterBearerToken
# Your bearer token. Application at developer.twitter.com required. You can store
# your bearer token in your .Rprofile in your home directory, creating following entry:
# twitterBearerToken <- "replace with your token"
query <- "#OccupyWallStreet"
startdate <- "2012-07-12T00:00:01Z"
enddate <- "2012-07-15T00:00:01Z"
occupy_search <- twitter_fas(query, startdate, enddate, bearer_token)
occupy_tweets <- occupy_search$tweets$tweets
hist(occupy_tweets$public_metrics_reply_count)
}
